requirements.txt
datasets
glob
json
matplotlib
numpy
pandas
requests
scikit-learn
scipy.cluster.hierarchy
scipy.spatial.distance
ta
tensorflow
torch
transformers
xgboost
yfinance
fpdf
dateutil.relativedelta